requirements.txt
alembic==1.4.3
anyio==3.0.0
asgiref==3.4.1
async-generator==1.10
click==8.0.3
contextvars==2.4
fastapi==0.70.0
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
immutables==0.16
importlib-metadata==4.8.3
importlib-resources==5.4.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2==2.9.3
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-editor==1.0.4
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.16.0
typing_extensions==4.0.1
uvicorn==0.16.0
zipp==3.6.0